convert to module step 1
